pyproject.toml
[tool.poetry]
name = "zen-app-backend"
version = "0.1.0"
description = "AI Backend for Zen APP"
authors = ["Trinh Quoc Trung <trinhquoctrung@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
python-dotenv = "^1.2.1"
pydantic-settings = "^2.12.0"
supabase = "^2.27.2"
postgrest = "^2.27.2"
asyncpg = "^0.31.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.46"}
pgvector = "^0.4.2"
openai = "^2.15.0"
tiktoken = "^0.12.0"
pyhumps = "^3.8.0"
sse-starlette = "^3.2.0"
python-multipart = "^0.0.22"
google-cloud-speech = "^2.36.1"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
black = "^26.1.0"
ruff = "^0.14.14"
mypy = "^1.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
